refactor: split test_integration into 5 feature test files, remove .feature docs
test/
  test_conversation.cpp  — chat, history, clear
  test_commands.cpp      — !, !!, /help, /version, unknown, empty, quit
  test_options.cpp       — /set, toggles, --no-color, --why-so-serious
  test_annotations.cpp   — write y/n/s, exec y/n
  test_markdown.cpp      — renderer, on/off toggle
  test_helpers.h         — shared MockLLM + test_cfg

Removed test/features/ — doctest SCENARIO is already Gherkin-style.
